+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/impl/TestProcedureDAOImpl.java
@@ -1,8 +1,9 @@
 package gov.healthit.chpl.dao.impl;
 
 import java.util.ArrayList;
-import java.util.Date;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import javax.persistence.Query;
 
@@ -10,15 +11,12 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
-import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.stereotype.Repository;
 
-import gov.healthit.chpl.auth.Util;
-import gov.healthit.chpl.dao.EntityCreationException;
-import gov.healthit.chpl.dao.EntityRetrievalException;
 import gov.healthit.chpl.dao.TestProcedureDAO;
+import gov.healthit.chpl.dto.TestProcedureCriteriaMapDTO;
 import gov.healthit.chpl.dto.TestProcedureDTO;
+import gov.healthit.chpl.entity.TestProcedureCriteriaMapEntity;
 import gov.healthit.chpl.entity.TestProcedureEntity;
 
 @Repository("testProcedureDAO")
@@ -26,95 +24,10 @@ public class TestProcedureDAOImpl extends BaseDAOImpl implements TestProcedureDA
     private static final Logger LOGGER = LogManager.getLogger(TestProcedureDAOImpl.class);
     @Autowired
     MessageSource messageSource;
-
+    
     @Override
-    public TestProcedureDTO create(TestProcedureDTO dto) throws EntityCreationException {
-
-        TestProcedureEntity entity = null;
-        if (dto.getId() != null) {
-            entity = this.getEntityById(dto.getId());
-        }
-
-        if (entity != null) {
-            throw new EntityCreationException("An entity with this ID already exists.");
-        } else {
-            entity = new TestProcedureEntity();
-            entity.setCreationDate(new Date());
-            entity.setDeleted(false);
-            entity.setLastModifiedDate(new Date());
-            entity.setLastModifiedUser(Util.getCurrentUser().getId());
-            entity.setVersion(dto.getVersion());
-
-            try {
-                create(entity);
-            } catch (Exception ex) {
-                String msg = String.format(messageSource.getMessage(
-                        new DefaultMessageSourceResolvable("listing.criteria.badTestProcedure"),
-                        LocaleContextHolder.getLocale()), dto.getVersion());
-                LOGGER.error(msg, ex);
-                throw new EntityCreationException(msg);
-            }
-            return new TestProcedureDTO(entity);
-        }
-    }
-
-    @Override
-    public TestProcedureDTO update(TestProcedureDTO dto) throws EntityRetrievalException {
-        TestProcedureEntity entity = this.getEntityById(dto.getId());
-
-        if (entity == null) {
-            throw new EntityRetrievalException("Entity with id " + dto.getId() + " does not exist");
-        }
-
-        entity.setVersion(dto.getVersion());
-        entity.setLastModifiedUser(Util.getCurrentUser().getId());
-        entity.setLastModifiedDate(new Date());
-
-        update(entity);
-        return new TestProcedureDTO(entity);
-    }
-
-    @Override
-    public void delete(Long id) throws EntityRetrievalException {
-
-        TestProcedureEntity toDelete = getEntityById(id);
-
-        if (toDelete != null) {
-            toDelete.setDeleted(true);
-            toDelete.setLastModifiedDate(new Date());
-            toDelete.setLastModifiedUser(Util.getCurrentUser().getId());
-            update(toDelete);
-        }
-    }
-
-    @Override
-    public TestProcedureDTO getById(Long id) {
-
-        TestProcedureDTO dto = null;
-        TestProcedureEntity entity = getEntityById(id);
-
-        if (entity != null) {
-            dto = new TestProcedureDTO(entity);
-        }
-        return dto;
-    }
-
-    @Override
-    public TestProcedureDTO getByName(String versionName) {
-
-        TestProcedureDTO dto = null;
-        List<TestProcedureEntity> entities = getEntitiesByVersion(versionName);
-
-        if (entities != null && entities.size() > 0) {
-            dto = new TestProcedureDTO(entities.get(0));
-        }
-        return dto;
-    }
-
-    @Override
-    public List<TestProcedureDTO> findAll() {
-
-        List<TestProcedureEntity> entities = getAllEntities();
+    public List<TestProcedureDTO> getByCriteriaNumber(String criteriaNumber) {
+        Set<TestProcedureEntity> entities = getTestProcedureByCertificationCriteria(criteriaNumber);
         List<TestProcedureDTO> dtos = new ArrayList<TestProcedureDTO>();
 
         for (TestProcedureEntity entity : entities) {
@@ -125,49 +38,77 @@ public class TestProcedureDAOImpl extends BaseDAOImpl implements TestProcedureDA
 
     }
 
-    private void create(TestProcedureEntity entity) {
-
-        entityManager.persist(entity);
-        entityManager.flush();
-
-    }
-
-    private void update(TestProcedureEntity entity) {
-
-        entityManager.merge(entity);
-        entityManager.flush();
-    }
-
-    private List<TestProcedureEntity> getAllEntities() {
-        return entityManager
-                .createQuery("from TestProcedureEntity where (NOT deleted = true) ", TestProcedureEntity.class)
-                .getResultList();
-    }
-
-    private TestProcedureEntity getEntityById(Long id) {
-
-        TestProcedureEntity entity = null;
-
-        Query query = entityManager.createQuery(
-                "from TestProcedureEntity where (NOT deleted = true) AND (id = :entityid) ", TestProcedureEntity.class);
-        query.setParameter("entityid", id);
-        List<TestProcedureEntity> result = query.getResultList();
-
-        if (result.size() > 0) {
-            entity = result.get(0);
+    @Override
+    public TestProcedureDTO getByCriteriaNumberAndValue(String criteriaNumber, String value) {
+        TestProcedureEntity entity = getTestProcedureByCertificationCriteriaAndValue(criteriaNumber, value);
+        if (entity == null) {
+            return null;
         }
-
-        return entity;
+        return new TestProcedureDTO(entity);
     }
 
-    private List<TestProcedureEntity> getEntitiesByVersion(String versionName) {
+    @Override
+    public List<TestProcedureCriteriaMapDTO> findAllWithMappedCriteria() {
 
-        TestProcedureEntity entity = null;
+        List<TestProcedureCriteriaMapEntity> entities = 
+                entityManager.createQuery("SELECT tpMap "
+                        + "FROM TestProcedureCriteriaMapEntity tpMap "
+                        + "LEFT JOIN FETCH tpMap.testProcedure tp "
+                        + "LEFT JOIN FETCH tpMap.certificationCriterion cce "
+                        + "LEFT JOIN FETCH cce.certificationEdition "
+                        + "WHERE tpMap.deleted <> true "
+                        + "AND tp.deleted <> true "
+                        , TestProcedureCriteriaMapEntity.class).getResultList();
+        List<TestProcedureCriteriaMapDTO> dtos = new ArrayList<TestProcedureCriteriaMapDTO>();
 
-        Query query = entityManager.createQuery(
-                "from TestProcedureEntity where " + "(NOT deleted = true) AND (UPPER(version) = :versionName) ",
-                TestProcedureEntity.class);
-        query.setParameter("versionName", versionName.toUpperCase());
-        return query.getResultList();
+        for (TestProcedureCriteriaMapEntity entity : entities) {
+            TestProcedureCriteriaMapDTO dto = new TestProcedureCriteriaMapDTO(entity);
+            dtos.add(dto);
+        }
+        return dtos;
+    }
+    private Set<TestProcedureEntity> getTestProcedureByCertificationCriteria(String criteriaNumber) {
+        Query query = entityManager.createQuery("SELECT tpMap "
+                + "FROM TestProcedureCriteriaMapEntity tpMap "
+                + "JOIN FETCH tpMap.testProcedure tp "
+                + "JOIN FETCH tpMap.certificationCriterion cce "
+                + "JOIN FETCH cce.certificationEdition "
+                + "WHERE tpMap.deleted <> true "
+                + "AND tp.deleted <> true "
+                + "AND (UPPER(cce.number) = :criteriaNumber)"
+                , TestProcedureCriteriaMapEntity.class);
+        query.setParameter("criteriaNumber", criteriaNumber.trim().toUpperCase());
+        List<TestProcedureCriteriaMapEntity> results = query.getResultList();
+        
+        Set<TestProcedureEntity> tps = new HashSet<TestProcedureEntity>();
+        for(TestProcedureCriteriaMapEntity result : results) {
+            tps.add(result.getTestProcedure());
+        }
+        return tps;
+    }
+
+    private TestProcedureEntity getTestProcedureByCertificationCriteriaAndValue(String criteriaNumber, String value) {
+        Query query = entityManager.createQuery("SELECT tpMap "
+                + "FROM TestProcedureCriteriaMapEntity tpMap "
+                + "JOIN FETCH tpMap.testProcedure tp "
+                + "JOIN FETCH tpMap.certificationCriterion cce "
+                + "JOIN FETCH cce.certificationEdition "
+                + "WHERE tpMap.deleted <> true "
+                + "AND tp.deleted <> true "
+                + "AND (UPPER(cce.number) = :criteriaNumber) "
+                + "AND (UPPER(tp.name) = :value)"
+                , TestProcedureCriteriaMapEntity.class);
+        query.setParameter("criteriaNumber", criteriaNumber.trim().toUpperCase());
+        query.setParameter("value", value.trim().toUpperCase());
+        
+        List<TestProcedureCriteriaMapEntity> results = query.getResultList();
+        if(results == null || results.size() == 0) {
+            return null;
+        }
+        List<TestProcedureEntity> tps = new ArrayList<TestProcedureEntity>();
+        for(TestProcedureCriteriaMapEntity result : results) {
+            tps.add(result.getTestProcedure());
+        }
+        return tps.get(0);
     }
 }