+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dto/PendingCertificationResultTestProcedureDTO.java
@@ -9,6 +9,8 @@ public class PendingCertificationResultTestProcedureDTO implements Serializable 
     private Long id;
     private Long pendingCertificationResultId;
     private Long testProcedureId;
+    private TestProcedureDTO testProcedure;
+    private String enteredName;
     private String version;
 
     public PendingCertificationResultTestProcedureDTO() {
@@ -18,7 +20,11 @@ public class PendingCertificationResultTestProcedureDTO implements Serializable 
         this.setId(entity.getId());
         this.setPendingCertificationResultId(entity.getPendingCertificationResultId());
         this.setTestProcedureId(entity.getTestProcedureId());
-        this.setVersion(entity.getTestProcedureVersion());
+        if(entity.getTestProcedure() != null) {
+            this.testProcedure = new TestProcedureDTO(entity.getTestProcedure());
+        }
+        this.enteredName = entity.getTestProcedureName();
+        this.setVersion(entity.getVersion());
     }
 
     public Long getId() {
@@ -51,5 +57,21 @@ public class PendingCertificationResultTestProcedureDTO implements Serializable 
 
     public void setVersion(final String version) {
         this.version = version;
+    }
+
+    public TestProcedureDTO getTestProcedure() {
+        return testProcedure;
+    }
+
+    public void setTestProcedure(TestProcedureDTO testProcedure) {
+        this.testProcedure = testProcedure;
+    }
+
+    public String getEnteredName() {
+        return enteredName;
+    }
+
+    public void setEnteredName(String enteredName) {
+        this.enteredName = enteredName;
     }
 }