+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/certifiedProduct/CertifiedProduct2015Validator.java
@@ -12,14 +12,18 @@ import org.springframework.stereotype.Component;
 
 import gov.healthit.chpl.dao.AccessibilityStandardDAO;
 import gov.healthit.chpl.dao.MacraMeasureDAO;
+import gov.healthit.chpl.dao.TestDataDAO;
 import gov.healthit.chpl.dao.TestFunctionalityDAO;
+import gov.healthit.chpl.dao.TestProcedureDAO;
 import gov.healthit.chpl.dao.TestToolDAO;
 import gov.healthit.chpl.dao.search.CertifiedProductSearchDAO;
 import gov.healthit.chpl.domain.CQMResultCertification;
 import gov.healthit.chpl.domain.CQMResultDetails;
 import gov.healthit.chpl.domain.CertificationCriterion;
 import gov.healthit.chpl.domain.CertificationResult;
+import gov.healthit.chpl.domain.CertificationResultTestData;
 import gov.healthit.chpl.domain.CertificationResultTestFunctionality;
+import gov.healthit.chpl.domain.CertificationResultTestProcedure;
 import gov.healthit.chpl.domain.CertificationResultTestTool;
 import gov.healthit.chpl.domain.CertifiedProduct;
 import gov.healthit.chpl.domain.CertifiedProductQmsStandard;
@@ -34,7 +38,9 @@ import gov.healthit.chpl.dto.CertifiedProductDetailsDTO;
 import gov.healthit.chpl.dto.MacraMeasureDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultMacraMeasureDTO;
+import gov.healthit.chpl.dto.PendingCertificationResultTestDataDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultTestFunctionalityDTO;
+import gov.healthit.chpl.dto.PendingCertificationResultTestProcedureDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultTestTaskDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultTestTaskParticipantDTO;
 import gov.healthit.chpl.dto.PendingCertificationResultTestToolDTO;
@@ -43,7 +49,9 @@ import gov.healthit.chpl.dto.PendingCertifiedProductQmsStandardDTO;
 import gov.healthit.chpl.dto.PendingCqmCertificationCriterionDTO;
 import gov.healthit.chpl.dto.PendingCqmCriterionDTO;
 import gov.healthit.chpl.dto.PendingTestTaskDTO;
+import gov.healthit.chpl.dto.TestDataDTO;
 import gov.healthit.chpl.dto.TestFunctionalityDTO;
+import gov.healthit.chpl.dto.TestProcedureDTO;
 import gov.healthit.chpl.dto.TestToolDTO;
 import gov.healthit.chpl.manager.CertifiedProductDetailsManager;
 import gov.healthit.chpl.util.CertificationResultRules;
@@ -96,6 +104,10 @@ public class CertifiedProduct2015Validator extends CertifiedProductValidatorImpl
 
     @Autowired
     TestToolDAO testToolDao;
+    @Autowired
+    TestProcedureDAO testProcDao;
+    @Autowired
+    TestDataDAO testDataDao;
     @Autowired
     TestFunctionalityDAO testFuncDao;
     @Autowired
@@ -715,6 +727,44 @@ public class CertifiedProduct2015Validator extends CertifiedProductValidatorImpl
                     }
                 }
 
+                if(certRules.hasCertOption(cert.getNumber(), CertificationResultRules.TEST_PROCEDURE)
+                        && cert.getTestProcedures() != null && cert.getTestProcedures().size() > 0) {
+                    for (PendingCertificationResultTestProcedureDTO crTestProc : cert.getTestProcedures()) {
+                        if (crTestProc.getTestProcedure() == null) {
+                            product.getErrorMessages().add("Certification " + cert.getNumber() +
+                                    " must provide a test procedure.");
+                        } else if(crTestProc.getTestProcedure() != null && crTestProc.getTestProcedure().getId() == null) {
+                            TestProcedureDTO foundTestProc = 
+                                    testProcDao.getByCriteriaNumberAndValue(cert.getNumber(), crTestProc.getTestProcedure().getName());
+                            if(foundTestProc == null || foundTestProc.getId() == null) {
+                                product.getErrorMessages().add("Certification " + cert.getNumber()
+                                + " contains an invalid test procedure name: '" + crTestProc.getTestProcedure().getName() + "'.");
+                            } else {
+                                crTestProc.getTestProcedure().setId(foundTestProc.getId());
+                            }
+                        }
+                    }
+                }
+                
+                if(certRules.hasCertOption(cert.getNumber(), CertificationResultRules.TEST_DATA)
+                        && cert.getTestData() != null && cert.getTestData().size() > 0) {
+                    for (PendingCertificationResultTestDataDTO crTestData : cert.getTestData()) {
+                        if (crTestData.getTestData() == null) {
+                            product.getErrorMessages().add("Certification " + cert.getNumber() +
+                                    " must provide a test data.");
+                        } else if(crTestData.getTestData() != null && crTestData.getTestData().getId() == null) {
+                            TestDataDTO foundTestData = 
+                                    testDataDao.getByCriteriaNumberAndValue(cert.getNumber(), crTestData.getTestData().getName());
+                            if(foundTestData == null || foundTestData.getId() == null) {
+                                product.getErrorMessages().add("Certification " + cert.getNumber()
+                                + " contains an invalid test data name: '" + crTestData.getTestData().getName() + "'.");
+                            } else {
+                                crTestData.getTestData().setId(foundTestData.getId());
+                            }
+                        }
+                    }
+                }
+                
                 if (certRules.hasCertOption(cert.getNumber(), CertificationResultRules.G1_MACRA)
                         && cert.getG1MacraMeasures() != null && cert.getG1MacraMeasures().size() > 0) {
                     for (PendingCertificationResultMacraMeasureDTO pendingMeasureMap : cert.getG1MacraMeasures()) {
@@ -1370,6 +1420,44 @@ public class CertifiedProduct2015Validator extends CertifiedProductValidatorImpl
                     }
                 }
 
+                if(certRules.hasCertOption(cert.getNumber(), CertificationResultRules.TEST_PROCEDURE)
+                        && cert.getTestProcedures() != null && cert.getTestProcedures().size() > 0) {
+                    for (CertificationResultTestProcedure crTestProc : cert.getTestProcedures()) {
+                        if (crTestProc.getTestProcedure() == null) {
+                            product.getErrorMessages().add("Certification " + cert.getNumber() +
+                                    " must provide a test procedure.");
+                        } else if(crTestProc.getTestProcedure() != null && crTestProc.getTestProcedure().getId() == null) {
+                            TestProcedureDTO foundTestProc = 
+                                    testProcDao.getByCriteriaNumberAndValue(cert.getNumber(), crTestProc.getTestProcedure().getName());
+                            if(foundTestProc == null || foundTestProc.getId() == null) {
+                                product.getErrorMessages().add("Certification " + cert.getNumber()
+                                + " contains an invalid test procedure name: '" + crTestProc.getTestProcedure().getName() + "'.");
+                            } else {
+                                crTestProc.getTestProcedure().setId(foundTestProc.getId());
+                            }
+                        }
+                    }
+                }
+                
+                if(certRules.hasCertOption(cert.getNumber(), CertificationResultRules.TEST_DATA)
+                        && cert.getTestDataUsed() != null && cert.getTestDataUsed().size() > 0) {
+                    for (CertificationResultTestData crTestData : cert.getTestDataUsed()) {
+                        if (crTestData.getTestData() == null) {
+                            product.getErrorMessages().add("Certification " + cert.getNumber() +
+                                    " must provide a test data.");
+                        } else if(crTestData.getTestData() != null && crTestData.getTestData().getId() == null) {
+                            TestDataDTO foundTestData = 
+                                    testDataDao.getByCriteriaNumberAndValue(cert.getNumber(), crTestData.getTestData().getName());
+                            if(foundTestData == null || foundTestData.getId() == null) {
+                                product.getErrorMessages().add("Certification " + cert.getNumber()
+                                + " contains an invalid test data name: '" + crTestData.getTestData().getName() + "'.");
+                            } else {
+                                crTestData.getTestData().setId(foundTestData.getId());
+                            }
+                        }
+                    }
+                }
+                
                 if (certRules.hasCertOption(cert.getNumber(), CertificationResultRules.G1_MACRA)
                         && cert.getG1MacraMeasures() != null && cert.getG1MacraMeasures().size() > 0) {
                     for (int i = 0; i < cert.getG1MacraMeasures().size(); i++) {